+++ corporate/src/main/resources/application-prod.yml
@@ -1,0 +1,40 @@
+server:
+  port: 19094
+
+spring:
+  application:
+    name: corporate
+
+  datasource:
+    url: jdbc:postgresql://localhost:5432/corpoate
+    username: crp
+    password: 1234
+    driver-class-name: org.postgresql.Driver
+
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+        format_sql: true
+  elasticsearch:
+    uris: ${ELASTICSEARCH-ENDPOINT}
+    username: elastic
+    password: ${ELASTICSEARCH-PASSWORD}
+  data:
+    redis:
+      host: localhost
+      port: 6379
+
+eureka:
+  client:
+    service-url:
+      defaultZone: http://localhost:19090/eureka/
+
+logging:
+  file:
+    name: corporate/src/main/logs/corporate.log
+  level:
+    root: INFO
